+++ src/apiIndex.ts
@@ -1,4 +1,4 @@
-// Copyright 2018, Google, LLC.
+// Copyright 2020 Google LLC
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -16,7 +16,7 @@ import {GoogleConfigurable, ServiceOptions} from '.';
 export function getAPI<T>(
   api: string,
   options: ServiceOptions | string,
-  // tslint:disable-next-line no-any
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   versions: {[index: string]: any},
   context?: GoogleConfigurable
 ) {