+++ synth.py
@@ -11,10 +11,15 @@
 
 import synthtool as s
 import synthtool.gcp as gcp
+import synthtool.languages.node as node
 import logging
-import subprocess
 
 logging.basicConfig(level=logging.DEBUG)
+
+AUTOSYNTH_MULTIPLE_COMMITS = True
+
 common_templates = gcp.CommonTemplates()
 templates = common_templates.node_library()
 s.copy(templates)
+node.install()
+node.fix()